refactor: move diagram methods to new class
